F04 - Fixed switch month
