+++ assets/js/_handlers.js
@@ -1,7 +1,7 @@
 import { swapTemplate, removeTemplate } from "./_templates.js";
 import { formValidation } from "./_form_validation.js";
-import { addMonth, currentDate } from "./calendar.js";
 import { calendarEvent } from "./_events.js";
+import * as render from "./_month_render.js";
 
 /*
  * All listeners are listed here
@@ -34,6 +34,8 @@ export function handleDocumentEvents(e) {
       if (!formValidation(e, true)) {
         const data = calendarEvent.getDataFromModal("#modal form");
         calendarEvent.toLocalStorage(data);
+        
+        console.log(calendarEvent.fromLocalStorage(data));
         // calendarEvent.probando();
       }
     }
@@ -42,13 +44,9 @@ export function handleDocumentEvents(e) {
      * buttons to switch month
      */
     if (e.target.matches(".fa-chevron-right")) {
-      let updatedMonth = currentDate.getMonth();
-      let updatedYear = currentDate.getFullYear();
       addMonth(updatedYear, updatedMonth, true);
     }
     if (e.target.matches(".fa-chevron-left")) {
-      let updatedMonth = currentDate.getMonth();
-      let updatedYear = currentDate.getFullYear();
       addMonth(updatedYear, updatedMonth, false);
     }
 
@@ -74,4 +72,15 @@ export function handleDocumentEvents(e) {
       formValidation(e, false);
     }
   });
+}
+
+let updatedMonth = (new Date()).getMonth();
+let updatedYear = (new Date()).getFullYear();
+function addMonth(year, month, boolean) {
+  boolean ? month++ : month--;
+  updatedYear = render.updateDate(year,month).year;
+  updatedMonth = render.updateDate(year,month).month;
+  swapTemplate("month","calendar");
+  render.renderMonth(updatedYear,updatedMonth);
+  render.addTag(updatedYear, updatedMonth);
 }
